+++ CHANGELOG.md
@@ -10,9 +10,18 @@ and this project adheres to
 
 ### Added
 
+- `mobileDeviceManagementAuthority`, `intuneSubscriptionState`, and
+  `intuneAccountId` properties to `intune_managed_device` entities
+
+### Changed
+
+- Only ingest Intune data if the Microsoft 365 account has Intune configured
+
+### Added
+
 - A class of `CompliancePolicy` to `intune_compliance_policy` and
   `intune_device_configuration` entities
-- `categpory` property to `intune_compliance_policy` and
+- `category` property to `intune_compliance_policy` and
   `intune_device_configuration` entities.
 
 - Support for ingesting the following **new** resources